1) Pull generic test code into its own file 2) Added two hacks into generic test to get over issue #2007 (under git repo sonic-utilities) 3) Moved some code to common 4) Added more test flags to run partial combos

+++ tests/configlet/util/run_test_in_switch.py
@@ -11,7 +11,8 @@ from helpers import *
 
 
 def run_test(skip_load=False, skip_clet_test=False,
-        skip_generic_add=False, skip_generic_rm=False):
+        skip_generic_add=False, skip_generic_rm=False,
+        skip_prepare=False):
     global data_dir, orig_db_dir, clet_db_dir, files_dir
 
     set_print()
@@ -29,7 +30,8 @@ def run_test(skip_load=False, skip_clet_test=False,
     do_test_add_rack(duthost, skip_load=skip_load,
             skip_clet_test=skip_clet_test,
             skip_generic_add=skip_generic_add,
-            skip_generic_rm=skip_generic_rm)
+            skip_generic_rm=skip_generic_rm,
+            skip_prepare=skip_prepare)
     
 
 def main():
@@ -38,6 +40,8 @@ def main():
 
     parser=argparse.ArgumentParser(description="Sample for argparse")
     parser.add_argument("-d", "--dir", help="Test Data dir", required=True)
+    parser.add_argument("-w", "--use-exist", help="Use data dir as working dir", 
+            action='store_true', default=False)
     parser.add_argument("-l", "--skip-load", help="skip initial minigraph loading", 
             action='store_true', default=False)
     parser.add_argument("-c", "--skip-clet-test",
@@ -56,14 +60,17 @@ def main():
         print("dir does not exist")
         return -1
 
-    work_dir = os.path.join(args.dir, datetime.datetime.now().strftime("%d_%m_%Y_%H_%M"))
-
-    try:
-        os.system("rm -rf {}".format(work_dir))
-        os.mkdir(work_dir)
-    except OSError as error:
-        print("Failed to create {} with err:{}".format(work_dir, str(error)))
-        return -1
+    if not args.use_exist:
+        work_dir = os.path.join(args.dir, datetime.datetime.now().strftime("%d_%m_%Y_%H_%M"))
+        try:
+            os.system("rm -rf {}".format(work_dir))
+            os.mkdir(work_dir)
+        except OSError as error:
+            print("Failed to create {} with err:{}".format(work_dir, str(error)))
+            return -1
+    else:
+        work_dir = args.dir
+        skip_load = True
 
     os.chdir(work_dir)
 
@@ -71,7 +78,8 @@ def main():
 
     run_test(skip_load=args.skip_load, skip_clet_test=args.skip_clet_test,
             skip_generic_add=args.skip_generic_add,
-            skip_generic_rm=args.skip_generic_rm)
+            skip_generic_rm=args.skip_generic_rm,
+            skip_prepare = args.use_exist)
 
 
 if __name__ == "__main__":
